rename: api 패키지 자료 구조 변경
